+++ examples/cmu_arctic_corpus.py
@@ -24,19 +24,16 @@ if __name__ == '__main__':
     args = parser.parse_args()
 
     # Download the corpus, be patient
-    corpus = pra.datasets.CMUArcticCorpus(download=True, speakers=['bdl'])
+    corpus = pra.datasets.CMUArcticCorpus(download=True, speaker=['bdl'])
 
-    # the corpus is split between train and test
-    # let's pick a sentence from each
-    print('A couple of sentences:')
-    print(corpus.sentences[0])
-    print(corpus.sentences[100])
+    # print dataset info and 10 sentences
+    corpus.head(n=10)
 
     # let's find all the sentences from male speakers in the training set
     keyword = 'what'
-    matches = list(filter(lambda x: keyword in x.text, corpus.sentences))
+    matches = corpus.filter(text=lambda t : keyword in t)
     print('The number of sentences containing "{}": {}'.format(keyword, len(matches)))
-    for s in matches:
+    for s in matches.sentences:
         print('  *', s)
 
     # play sound if required